Add support for setting up VM agents in an azure load balancer backend pool
Based heavily on work by @gpolek (https://github.com/jenkinsci/azure-vm-agents-plugin/pull/61)

Allows the specification of a load balancer & backend (that can optionally be
in a different resource group to that in which the Agent VMs get spun up)

This is (as mentioned in the original pull request) highly useful for scenarios
where you need to whitelist the traffic coming from your pool of VMs.

This is a pre-requisite for support AAD logins to the jenkins agents too on
environments where Conditional Access Policies use named location constraints.