Spare Pool & Simultaneous Identity
The two main features of this PR are:
1) Spare Pool: this is an option for the pool
retention strategy that allows the pool size
parameter to specify how many spare instances
should be availble. This allows a small pool
to be maintained during low usage times, while
ensuring that there are always ready nodes
for incoming builds.

2) Single Use Pool: This is an option for the pool
retention strategy to only use each node once.

3) Allow System Assigned and  User assigned identities
to be allocated simultaneously. The current implementation
is checkboxes which makes it look like both can be
requested, but the code only allows either/or.

4) Raise minimum Jenkins version to 2.150.3

5) Increase frequency of maintenance tasks so that nodes
are cleaned up efficiently. On a busy master, having nodes
waiting for minutes to be cleaned up means pool sizes have
to be higher to account for nodes waiting to be removed.

Related: conjurinc/ops#553